Fix errors in gcc-4.9 on Mac OS X
Statically cast the macros to int (from chtype)
Include <string> in globals.cpp

+++ src/globals.cpp
@@ -1,4 +1,5 @@
 #include "globals.h"
+#include <string>
 
 int TOTAL_POINTS = 0;
 int GAME_WON = 0;
@@ -20,6 +21,6 @@ int CURRENT_LEVEL = 0;
 //int walllist[] = {'#', ACS_ULCORNER, ACS_LLCORNER, ACS_URCORNER,
 //		ACS_LRCORNER, ACS_LTEE, ACS_RTEE, ACS_BTEE, ACS_TTEE,
 //		ACS_HLINE, ACS_VLINE, ACS_PLUS};
-int zz[] = {'#', 4194144, ACS_PLUS, 'b', 'c'};
+int zz[] = {'#', 4194144, static_cast<int>(ACS_PLUS), 'b', 'c'};
 //std::set<chtype> WALLS(zz, zz+5);
 std::mutex mtx;
